MultiAgents/MA/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.10.0
google-generativeai==0.3.2
supabase==2.3.4
pydantic==2.5.3
httpx==0.26.0
